fix OtherIncome conflicts
