+++ app/views/OtherIncomeFromPropertyView.scala.html
@@ -54,7 +54,7 @@
 
         @govukInput(
             InputViewModel(
-                field = form("otherIncomeFromProperty"),
+                field = form("amount"),
                 label = LabelViewModel(messages(s"otherIncomeFromProperty.para2.${individualOrAgent}"))
             )
             .withCssClass("govuk-fieldset__legend govuk-label--m")
